Move token parser to separate module
